First commit to the new-web branch. Lots of changes
I started by copying in the files from 'contrib/exfoliation' which has been
the default UI for a while. That includes the stylesheets, images, and all
the stuff in 'exfoliation/images/logos' that wasn't in 'html/images/logos'.

Then I worked on changing the stylesheets some more. Some changes to
default settings, fonts. Fixed errors in cmd.css and changed a few things.
common.css - there was a lot of stuff in there that isn't common, but
unique to main.php or the nav bar, so those got moved out; added a
shadow-box class and some infobox classes; fixed some errors.
status.css - fixed some errors and made a few changes.
tac.css - Quite a bit changed in there. See the entry on 'tac.c' below.

cgi/cgiutils.c - changed the display_info_table() function to use the new
infobox classes.

cgi/cmd.c - changed the display of the top (info box and context help stuff),
cleaned up the display for commands with no options, changed to look of
the command description and the "Please enter all" box at the bottom.

cgi/status.c - changed the display of the top (info box) and added a 'linkBox'
class for the links just below the info box. Changed how the totals boxes
at the top are formatted and look, including using the new shadowBox class.
Changed the detail hosts and services tables. More work to be done on the host
and service group displays and the network outages display.

cgi/tac.c - Made quite a few changes here, mostly to the look & feel, but also
some clean-up. In the Monitoring Features section, the checkmark/X has been
moved up into the column header. Clicking on the checkmark/X for Flap Detection,
Notifications, and Event Handlers now submits the enable/disable request via
an ajax call, instead of going to the cmd.cgi page.

The nav bar was moved from 'side.php' into 'index.php'. Frames are no longer
used, but the main display is now in an iframe. Changed the appearance of
the nav bar quite a bit.

Put in the current jquery (v3.1.0) to replace the really old v1.7.1.

Added a html/js/nagjs.js file for common javascript stuff. Currently contains
a popup window, a function to get the value of a cookie, and ajax functions.

Added a html/login.html page and changed the apache config so the website will
now use the login page instead of the ugly basic authentication window. There
are still a couple of issues that need to be addressed.

html/main.php - changed the appearance of the top and bottom sections somewhat.
Added a github logo/link next to the sourceforge one at the bottom.

Much more to come!